Remove dup code.
